+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 ### Changed
+- `JoinToLocation` is cacheable only if the joined query is also cacheable.
 
 ### Fixed